meme_generator/services/model_trainer.py
from meme_generator.models import UserImage
from meme_generator.database import db

from PIL import Image
import io

def fine_tune_model():
    # Placeholder for fine-tuning logic
    images_data = UserImage.query.all()
    fine_tuning_data = []

    for data in images_data:
        img = Image.open(io.BytesIO(data.image))
        fine_tuning_data.append({"image": img, "text": data.prompt})

    # Fine-tuning logic would be implemented here, potentially uploading to Hugging Face Hub for training
    return "Fine-tuning logic started"
